Added User-Location association. Added Set<UserEntity> in Role with LAZY fetch. Added check of associated entities before deleting Role, Location, Category with new method find With Associated*() in their Services. Fetch is LAZY now in Set<Discount Entity> in Category and Location. In Discount Entity there is no now default constructor. Categories and Area field of Discount Entity is now @JsonIgnore. All Entities tests passed.

+++ src/main/java/com/andersenlab/benefits/controller/CategoryController.java
@@ -13,6 +13,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 /**
  * The controller for handling requests for {@link CategoryEntity}.
@@ -117,6 +118,9 @@ public class CategoryController {
     public void deleteCategory(@PathVariable final Long id) {
         categoryService.findById(id).orElseThrow(() ->
                 new IllegalStateException("Category with id: '" + id + "' was not found in the database"));
+        final Optional<CategoryEntity> categoryEntity = categoryService.findWithAssociatedDiscounts(id);
+        if (categoryEntity.isPresent() && categoryEntity.get().getDiscounts().size() > 0)
+            throw new IllegalStateException("There is active discounts in this Category in database");
         categoryService.delete(id);
     }
