+++ src/main/java/com/andersenlab/benefits/controller/LocationController.java
@@ -145,6 +145,9 @@ public class LocationController {
     public void deleteLocation(@PathVariable final Long id) {
         locationService.findById(id).orElseThrow(() ->
                 new IllegalStateException("Location with id: '" + id + "' was not found in the database"));
+        final Optional<LocationEntity> locationEntity = locationService.findWithAssociatedDiscounts(id);
+        if (locationEntity.isPresent() && locationEntity.get().getDiscounts().size() > 0)
+            throw new IllegalStateException("There is active discounts in this Location in database");
         locationService.delete(id);
     }
 
@@ -181,8 +184,8 @@ public class LocationController {
                     description = "Internal Server Error",
                     content = @Content)
     })
-    @RequestMapping(method = RequestMethod.GET, value = "/locations/{country}", params = {"country"})
-    List<Optional<LocationEntity>> findByCountry(@RequestParam final String country) {
+    @RequestMapping(method = RequestMethod.GET, value = "/locations/country")
+    public List<Optional<LocationEntity>> findByCountry(@RequestParam final String country) {
         return locationService.findByCountry(country);
     }
 
@@ -200,9 +203,9 @@ public class LocationController {
                     description = "Internal Server Error",
                     content = @Content)
     })
-    @RequestMapping(method = RequestMethod.GET, value = "/locations/{country}/{filterMask}", params = {"country", "filterMask"})
-    List<Optional<LocationEntity>> findByFirstLetters(@RequestParam final String country,
-                                                      @RequestParam final String filterMask) {
+    @RequestMapping(method = RequestMethod.GET, value = "/locations/filter", params = {"country", "filterMask"})
+    public List<Optional<LocationEntity>> findByFirstLetters(@RequestParam final String country,
+                                                             @RequestParam final String filterMask) {
         return locationService.findByFirstLetters(country, filterMask);
     }
 }
